requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.1
psycopg2==2.9.9
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
tzlocal==5.2
urllib3==2.0.7
yarl==1.9.2
